requirements.txt
rich>=10.0.0
python-dotenv>=1.0.0
openai>=1.3.0 